[Fix] Autoscaling: Use pointers for structs
